[web] Updated all Gestures
