+++ web/Errors.js
@@ -1,0 +1,5 @@
+export class GesturePropError extends Error {
+  constructor(name, value, expectedType) {
+    super(`Invalid property \`${name}: ${value}\` expected \`${expectedType}\``);
+  }
+}
