+++ web/PanGestureHandler.js
@@ -1,0 +1,342 @@
+import Hammer from 'hammerjs';
+
+import {
+  MULTI_FINGER_PAN_MAX_PINCH_THRESHOLD,
+  MULTI_FINGER_PAN_MAX_ROTATION_THRESHOLD,
+} from './constants';
+import GestureHandler from './GestureHandler';
+import { isnan, TEST_MIN_IF_NOT_NAN, VEC_LEN_SQ } from './utils';
+
+class PanGestureHandler extends GestureHandler {
+  get name() {
+    return 'pan';
+  }
+
+  get shouldEnableGestureOnSetup() {
+    return true;
+  }
+
+  _getHammerConfig() {
+    return {
+      ...super._getHammerConfig(),
+      direction: this.getDirection(),
+    };
+  }
+
+  getDirection() {
+    const config = this.getConfig();
+    const {
+      activeOffsetXStart,
+      activeOffsetXEnd,
+      activeOffsetYStart,
+      activeOffsetYEnd,
+      minDist,
+    } = config;
+    let directions = [];
+    let horizontalDirections = [];
+
+    // sigh
+    if (!isnan(minDist)) {
+      return Hammer.DIRECTION_ALL;
+    }
+
+    if (!isnan(activeOffsetXStart)) horizontalDirections.push(Hammer.DIRECTION_LEFT);
+    if (!isnan(activeOffsetXEnd)) horizontalDirections.push(Hammer.DIRECTION_RIGHT);
+    if (horizontalDirections.length === 2) horizontalDirections = [Hammer.DIRECTION_HORIZONTAL];
+
+    directions = directions.concat(horizontalDirections);
+    let verticalDirections = [];
+
+    if (!isnan(activeOffsetYStart)) verticalDirections.push(Hammer.DIRECTION_UP);
+    if (!isnan(activeOffsetYEnd)) verticalDirections.push(Hammer.DIRECTION_DOWN);
+
+    if (verticalDirections.length === 2) verticalDirections = [Hammer.DIRECTION_VERTICAL];
+
+    directions = directions.concat(verticalDirections);
+
+    if (!directions.length) {
+      return Hammer.DIRECTION_NONE;
+    }
+    if (
+      directions[0] === Hammer.DIRECTION_HORIZONTAL &&
+      directions[1] === Hammer.DIRECTION_VERTICAL
+    ) {
+      return Hammer.DIRECTION_ALL;
+    }
+    if (horizontalDirections.length && verticalDirections.length) {
+      return Hammer.DIRECTION_ALL;
+    }
+
+    return directions[0];
+  }
+
+  getConfig() {
+    if (!this._hasCustomActivationCriteria) {
+      // Default config
+      // If no params have been defined then this config should emulate the native gesture as closely as possible.
+      return {
+        minDistSq: 10,
+      };
+    }
+    return this.config;
+  }
+
+  shouldFailUnderCustomCriteria(
+    { deltaX, deltaY },
+    { failOffsetXStart, failOffsetXEnd, failOffsetYStart, failOffsetYEnd }
+  ) {
+    return (
+      (!isnan(failOffsetXStart) && deltaX < failOffsetXStart) ||
+      (!isnan(failOffsetXEnd) && deltaX > failOffsetXEnd) ||
+      (!isnan(failOffsetYStart) && deltaY < failOffsetYStart) ||
+      (!isnan(failOffsetYEnd) && deltaY > failOffsetYEnd)
+    );
+  }
+
+  shouldActivateUnderCustomCriteria(
+    { deltaX, deltaY, velocity },
+    {
+      activeOffsetXStart,
+      activeOffsetXEnd,
+      activeOffsetYStart,
+      activeOffsetYEnd,
+      minDistSq,
+      minVelocityX,
+      minVelocityY,
+      minVelocitySq,
+    }
+  ) {
+    if (!isnan(activeOffsetXStart) && deltaX < activeOffsetXStart) {
+      return true;
+    }
+    if (!isnan(activeOffsetXEnd) && deltaX > activeOffsetXEnd) {
+      return true;
+    }
+    if (!isnan(activeOffsetYStart) && deltaY < activeOffsetYStart) {
+      return true;
+    }
+    if (!isnan(activeOffsetYEnd) && deltaY > activeOffsetYEnd) {
+      return true;
+    }
+
+    if (TEST_MIN_IF_NOT_NAN(VEC_LEN_SQ({ x: deltaX, y: deltaY }), minDistSq)) {
+      return true;
+    }
+
+    if (TEST_MIN_IF_NOT_NAN(velocity.x, minVelocityX)) {
+      return true;
+    }
+    if (TEST_MIN_IF_NOT_NAN(velocity.y, minVelocityY)) {
+      return true;
+    }
+    if (TEST_MIN_IF_NOT_NAN(VEC_LEN_SQ(velocity), minVelocitySq)) {
+      return true;
+    }
+
+    return false;
+  }
+
+  _shouldMultiFingerPanFail({ pointerLength, scale, deltaRotation }) {
+    if (pointerLength <= 1) {
+      return false;
+    }
+
+    // Test if the pan had too much pinching or rotating.
+    const deltaScale = Math.abs(scale - 1);
+    const absDeltaRotation = Math.abs(deltaRotation);
+    if (deltaScale > MULTI_FINGER_PAN_MAX_PINCH_THRESHOLD) {
+      // > If the threshold doesn't seem right.
+      // You can log the value which it failed at here:
+      // console.log('Pan failed: scale: ', deltaScale);
+      return true;
+    }
+    if (absDeltaRotation > MULTI_FINGER_PAN_MAX_ROTATION_THRESHOLD) {
+      // > If the threshold doesn't seem right.
+      // You can log the value which it failed at here:
+      // console.log('Pan failed: rotate: ', absDeltaRotation);
+      return true;
+    }
+
+    return false;
+  }
+
+  updateHasCustomActivationCriteria({
+    minDistSq,
+    minVelocityX,
+    activeOffsetXStart,
+    activeOffsetYStart,
+    minVelocityY,
+    activeOffsetXEnd,
+    activeOffsetYEnd,
+    minVelocitySq,
+  }) {
+    return (
+      !isnan(minDistSq) ||
+      !isnan(minVelocityX) ||
+      !isnan(minVelocityY) ||
+      !isnan(minVelocitySq) ||
+      !isnan(activeOffsetXStart) ||
+      !isnan(activeOffsetXEnd) ||
+      !isnan(activeOffsetYStart) ||
+      !isnan(activeOffsetYEnd)
+    );
+  }
+
+  updateGestureConfig({
+    minVelocity = Number.NaN,
+    minVelocityX = Number.NaN,
+    minVelocityY = Number.NaN,
+    minPointers = Number.NaN,
+    maxPointers = Number.NaN,
+    minDist = Number.NaN,
+    activeOffsetXStart = Number.NaN,
+    activeOffsetXEnd = Number.NaN,
+    failOffsetXStart = Number.NaN,
+    failOffsetXEnd = Number.NaN,
+    activeOffsetYStart = Number.NaN,
+    activeOffsetYEnd = Number.NaN,
+    failOffsetYStart = Number.NaN,
+    failOffsetYEnd = Number.NaN,
+
+    ...props
+  }) {
+    // this.validateConfig(this.config)
+    return super.updateGestureConfig({
+      minVelocity,
+      minVelocityX,
+      minVelocityY,
+      minPointers,
+      maxPointers,
+      minDist,
+      activeOffsetXStart,
+      activeOffsetXEnd,
+      failOffsetXStart,
+      failOffsetXEnd,
+      activeOffsetYStart,
+      activeOffsetYEnd,
+      failOffsetYStart,
+      failOffsetYEnd,
+
+      ...props,
+    });
+  }
+
+  validateConfig(config = {}) {
+    const isNum = v => isnan(v) || typeof v === 'number';
+    const isBool = v => typeof v === 'boolean';
+
+    const valid = {
+      enabled: isBool,
+      minDistSq: isNum,
+      minVelocityX: isNum,
+      minVelocityY: isNum,
+      // TODO: Bacon: remove `minVelocity`
+      minVelocity: isNum,
+      minVelocitySq: isNum,
+      activeOffsetXStart: isNum,
+      activeOffsetXEnd: isNum,
+      failOffsetXStart: isNum,
+      failOffsetXEnd: isNum,
+      activeOffsetYStart: isNum,
+      activeOffsetYEnd: isNum,
+      failOffsetYStart: isNum,
+      failOffsetYEnd: isNum,
+      hasCustomActivationCriteria: isBool,
+      minPointers: isNum,
+      maxPointers: isNum,
+    };
+    const keys = Object.keys(valid);
+
+    let invalidKeys = [];
+    for (const key of Object.keys(config)) {
+      if (keys.includes(key)) {
+        if (valid[key](config[key])) {
+          console.warn('Invalid type: ' + key + ': ' + config[key]);
+        }
+      } else {
+        invalidKeys.push(key);
+      }
+    }
+
+    if (invalidKeys.length) {
+      throw new Error('Invalid config props found: ' + invalidKeys.join(', '));
+    }
+    return config;
+  }
+
+  isGestureEnabledForEvent(props, recognizer, inputData) {
+    if (this.shouldFailUnderCustomCriteria(inputData, props)) {
+      return { failed: true };
+    }
+
+    const velocity = { x: inputData.velocityX, y: inputData.velocityY };
+    if (
+      this._hasCustomActivationCriteria &&
+      this.shouldActivateUnderCustomCriteria(
+        { deltaX: inputData.deltaX, deltaY: inputData.deltaY, velocity },
+        props
+      )
+    ) {
+      if (
+        this._shouldMultiFingerPanFail({
+          pointerLength: inputData.maxPointers,
+          scale: inputData.scale,
+          deltaRotation: inputData.deltaRotation,
+        })
+      ) {
+        return {
+          failed: true,
+        };
+      }
+      return { success: true };
+    }
+
+    return { success: false };
+    // return validateCriteria(
+    //   {
+    //     ...inputData,
+    //     recognizer,
+    //     pointerLength: inputData.maxPointers,
+    //     velocity: inputData.overallVelocity,
+    //     vx: inputData.velocityX,
+    //     vy: inputData.velocityY,
+    //     dx: inputData.deltaX,
+    //     dy: inputData.deltaY,
+    //   },
+    //   props
+    // );
+  }
+
+  // The event object that is returned
+  parseNativeEvent({
+    translationX,
+    translationY,
+    velocityX,
+    velocityY,
+    x,
+    y,
+    absoluteX,
+    absoluteY,
+  }) {
+    return {
+      translationX,
+      translationY,
+      velocityX,
+      velocityY,
+      x,
+      y,
+      absoluteX,
+      absoluteY,
+    };
+  }
+
+  createNativeGesture({ manager, props }) {
+    manager.add(
+      new Hammer.Pan({
+        pointers: props.minPointers,
+      })
+    );
+  }
+}
+
+export default PanGestureHandler;
